DataKanjidic.rb: A script that provides an interface to the KANJIDIC format.
build-single-reading-kanji.rb: A script that produces a list of kanji with only one (non-nanori),
ordered by grade and then by Heisig index, in a format suitable for inclusion in an HTML table.

kanji-with-a-single-reading.jhtml: Updated in light of the output from the above script.

Note that these scripts are not used to automatically generate anything. Rather,
kanji-with-a-single-reading.jhtml has been manually updated with the output.