The first kanji keyword in the file is expected to be the one that will match the Heisig keyword. Adjust the code to complain if more than one kanji uses any given keyword. However, only complain if the kanji which repeats the keyword is within the first 2042 kanji that make up the RTK1 set.
Implementing the above exposed a number of kanji where the KANJIDIC data doesn't
exactly match the Heisig keyword (or where at least the first listed keyword doesn't
match Heisig). Adjust all the data exposed by the code. This doesn't guarantee that
all the data now matches the Heisig keywords but it does guarantee that there are
no duplicates within the RTK1 set *and* that kanji within the RTK1 set will be accessible
via their first-specified keyword.

japanese-to-codes.rb: If a repeated main kanji keyword is seen, complain to stderr. Do not complain
if the later kanji is from outside the RTK1 set. When multiple kanji have the same keyword only
accept the first one seen (which will be the numerically lowest one).

kanji.data: Adjust the kanji data in a few cases to match the keywords on http://kanji.koohii.com.