The recent problem with the "bull's eye" keyword suggests that it would be worthwhile checking that all of the RTK1 kanji can be geneated.
To this end ...

build-heisig-page.rb: ... produce a script that will generate a JHTML page that includes
an attempt to generate every kanji through @KJ{{keyword}} ...

Makefile: ... and cause the intermediate JHTML file and the final HTML　page to be generated.

KanjiSupport.rb: That leads to the realisation that the characters "(", ")" and "/" need to
be accounted for. In addition, the digits 0-9 should also be allowed in a keyword.

study-material.jhtml:　Having built the test page so that it can be grepped for an error message
(with the word UNKNOWN　in it), a link to it may as well be provided.