Build some pages with N1, N2 and N3 kanji vocabulary; build each of these both as a list of new vocabulary at that level and also as a cumulative list up to that level.
The vocabulary lists originally came from http://http://www.tanos.co.uk/jlpt/skills/vocab/ as .DOC files.
These were turned into .mht (HTML) files using MS Word. Those are fed as inputs to the build process, which uses
a ruby script to build a table of kanji with kana and English translation as a tool tip.

Note that the final output puts each word into one of three categories:
 - completely known: This word uses no kanji beyond the specified level
 - partly known: This word uses some kanji that are expected to be known at the specified level and some that are not.
 - unknown: None of the kanji used in this word are expected to be known at the specified level.

Ultimately the level of each kanji comes from the KANJIDIC file. This rates each kanji according to the pre-2010 JLPT
levels: 1-4. There is no unique classification for N3, so, erring on the side of safety, it is assumed that all
N2 kanji will be known at this level.

The original N4 vocabulary list was produced by a similar process some time ago, however the build process was
not committed to the repository. Rather than try to recreate this process, the original generated HTML file
is committed here and the Makefile modified to simply copy it into the output directory. This way a complete
rebuild in a freshly cloned repository should be complete.

external/VocabList.N1.mht:
external/VocabList.N2.mht:
external/VocabList.N3.mht: The original "new vocabulary at a given level" files.

external/VocabList.C.N1toN5.mht:
external/VocabList.C.N2toN5.mht:
external/VocabList.C.N3toN5.mht: The original "cumulative vocabulary up to a given level" files.

build-kanji-vocab-list-by-jlpt.rb: This script turns an MHT file into an HTML file with three sections according to how much
of the word is expected to be known at a given level.

N4-vocabulary-with-annotated-kanji.html: This is a similar file for N4 vocabulary which was originally generated. It is
being committed here to avoid having to recreate the process that built it.

index.jhtml: Add links for the new HTML pages. Remove some TODO cruft (all of which has now, in fact, been done).

Makefile: Build as required. It should be possible to generate the rules for building the new pages, but there are so
few of them that simply writing them out is good enough for now.