Change japanese-to-codes.rb to use a different processing method. The advantage of the new method is that it allows codes to be embedded within one another. There is no change in the output produced (although one or two hacks have been left in to achieve this: these will be removed in the next commit).
The new script could do with some tidying up and some documenting, but that is for
a future commit.

The new script produces identical output to the old one and takes almost exactly
as long as the old one to process the current data (14m8s for the old one and
14m14s for the new one on the Raspberry Pi running ruby 1.9.3). On a powerful server
running 1.8.7 the new script is marginally faster.

Rename japanese-to-codes.rb -> japanese-to-codes.old.rb

japanese-to-codes.rb: The new processing script.

Makefile: Update the makefile to account for the script output change.