Add @GRMIDX{{...}} support.
Begin to improve the commenting.

Add @GRMIDX{{}} support. The intention is that @GRMIDX{{}} will be processed to produce an automatic index. However
the default processing will be to drop it entirely, which is what should happen when processing a JHTML file.

AtCommandSupport.rb: Improve comments and accept @GRMIDX{{...}} but produce no output for it.