First cut at building an automatically maintained grammar index.
Build an index of grammatical points so that the woefully out of date manually maintained one can eventually be retired.

find-grammar-elements.rb: A script that examines a JHTML file and pulls out @GRMIDX{{...}} statements.

build-grammar-index.rb: A script that takes a set of @GRMIDX{{...}} statements spread over a number of files
and builds a grammar page based on the contents.

Makefile: Process a set of JHTML files to produce GRMIDX files. Then process all of these together to
build an automatically maintained index.

.gitignore: Ignore the generated/ subdirectory.