Add support for furigana with @FG{{}}.
AtCommandSupport.rb: Add @FG{{}}. Process by splitting on the first ":". Everything before
will be processed as hiragana and displayed. Everything afterwards will be processed as hiragana
and rendered as a tooltip.

DebugSupport.rb: Add support for diagnostic_out(). This is displays to stderr but, unlike debug_out, is displayed
even if debugging is disabled.