Makefile: add some sources that exist and should be built.
Running Xenu against bin/index.html found a number of broken links. Some of these were links in doijg.jhtml
that were for pages that will eventually be written but do not currently exist. These have been fixed to
point to "grammar-.html".

Other broken links were to pages for which source exists but happened not to be built. This change causes
those sources to be accounedt for in the makefile.