Fix bad use of <span> ... </span> to format text.
A long time ago the adjectives page used <span> in a table to cause various endings to be
highlighted. That highlighting changed quite a while ago, perhaps unintentionally because
the CSS "color:" directive was missing a "#" to denote a hexadecimal number. The "font-weight:"
directive, however, still had an effect which caused all spans within a "sentence" table to be rendered
in bold. That unfortunately affects the span that @FG{{}} uses to provide furigana, causing the
base text to appear in bold.

The unqualified span entry in _sentence.scss (and the one in _example.scss) only have the effect
of making text within any span bold, because the "color:" directive is badly formatted. Replace
the uses of these spans in adjectives.html with the more reasonable @EM{{}} and give the spans names
so that the formatting only applies where it is required (in future).

adjectives.jhtml: Instead of marking emphasis with <span> ... </span>, use @EM{{}}.

_example.scss:
_sentence.scss: Fix the "color:" entry to include "#" for a hex number and give the (unused) span a name.