Move ruby scripts to scripts/ directory and data to data/ directory.
Begin the process of making the build process more manageable.
Put all scripts in a scripts/ directory.
Put locally created data into a data/ directory.

scripts/build-heisig-page.rb: This script has the name of kanji.data hardcoded so that needs to change.

scripts/RefsSupport.rb: This script has the name of references.data hardcoded so that needs to change.

Makefile: Minimal changes to cope with the relocation of ruby scripts and data files.

00ideas.txt: Random ideas. Nothing really changed here but it's easier to commit this along with the rest.