Improve the pages for "amari" and "ba ... hodo" grammar points.
grammar/grammar-amari.jhtml: Add "anmari" as a grammar point index. Improve the formation and text to account for the use of
"amari" as a conjunction. Add DOIJG as a reference.

grammar/grammar-ba-hodo.jhtml: Expand the formation. Move the example closer to the relevant text.
Add a paragraph to cover the use of "ba hodo" with "suru" verbs. Add a reference to DOIJG.

grammar/grammar-verb-use-of-passive.jhtml: Try to fix the use of <dt> tags.

study-material/progress/dobjg.jhtml: Put in the correct anchor for the "amari" reference.

study-material/progress/doijg.jhtml: Sort out the entries for "amari" and "ba ... hodo".