Fix navbars to not place the target behind the navigation bar.
The first part of the fix is to move the multiple identical copies of the navgiation bar code
into _nav.scss, so that the needless duplication is removed.

Then all the adhoc <p></p> code was removed: it wasn't doping any good.

The most obvious problem with NSM nav bar elements was that they were being limited to a eidth of 60 pixels.
Instead provide the required spacing by adding some padding.

That leaves just the fact that the navigation bar was causing the selected target to appear at the top of the page
rather than at the top of its ,div>, i.e. below the nav bar. The trick to making this work is to set the target
up to be enough pixels below the top of the page to not be behind the nav bar. That arbitrary pixel margin may
not be a general solution but it works for all of the current TOC pages.

The final problem was that once a target was selected, the first few navigation items became unavailable, basically
all the ones behind which the anchor text (e.. "Chapter 1") lay. The fix is to only anchor on an initial underscore
and to make that character's visibility hidden.

css/_nav.scss: New navbar code.

css/japanese.scss: Include new navbar code.

study-material/toc/kzm-n2g-toc.jhtml:
study-material/toc/kzm-n3g-toc.jhtml:
study-material/toc/mnn-int-toc.jhtml:
study-material/toc/mnn-toc.jhtml:
study-material/toc/nsm-n2g-toc.jhtml:
study-material/toc/nsm-n3g-toc.jhtml: Remove redundant copies of the navigation bar code. Fix the anchors so that they will work as intended.