It is about time that kanji.data was verified against the RTK website. Provide a script to do that.
Actually this isn't verifying against the website, but against a data file that purports to list the keywords Heisig used. The script assumes that you are using RTK 5th edition keywords. It needs to be manually run.

The two issues the script spots are:
  "sheet of" but the RTK website uses "a sheet of"
and
  "brush stroke" but the RTK website uses "brush-stroke"

data/heisig-data.txt: Master data file of heisig keywords (for 5th edition).

scripts/check-kanji-against-heisig.rb: verification script.