Take the first step in eliminating the old, hand-maintained grammar.jhtml.
Work through grammar.jhtml and eliminate everything that is in the automatically-generated grammar-index.html.

In the process a number of missing GRMIDX entries needed to be added and a few similar issues needed to be resolved.

grammar/Makefile: grammar-ya-store was not being built ...

grammar/grammar-ya-store.jhtml: ... and the index text could be clearer. In passing, move the example nearer to the explanatory text.

grammar/grammar-adverbs-of-degree.jhtml:
grammar/grammar-basic-sentence-patterns.jhtml:
grammar/grammar-expressing-time.jhtml:
grammar/grammar-goro.jhtml:
grammar/grammar-iku.jhtml:
grammar/grammar-interrogatives-with-particles.jhtml:
grammar/grammar-polite-and-plain.jhtml: Add a missing GRMIDX.

grammar/grammar-tsukuru.jhtml: Try to clarify the existing GRMIDX.

grammar/grammar.jhtml: Remove all of the hand-maintained code, except for the two entries that currently cannot have GRMIDX entries.

grammar/grammar-kuru.jhtml: Add a missing GRMIDX that was noticed in passing. This one was not in the hand-maintained code either.