If a build fails, a subsequent build should not pass without some editing.
If a makerule fails, the output file should be removed so that a subsequent make does not find it and think that nothing needs to be done. This requires that fatal build failures always generate a non-zero exit code.

makerules/rules.make: If building from a JHTML source file fails, remove the partial output file.

scripts/HiraganaSupport.rb:
scripts/find-grammar-elements.rb: When exiting on failure, specify a non-zero code.

scripts/KanjiSupport.rb:
scripts/KatakanaSupport.rb: When exiting on failure, specify a non-zero code and issue the error message on STDERR.