More detailed updates from KZM-N2G Chpater 1.1.
grammar/grammar-ka-to-omou-to.jhtml: Fix reference.

grammar/grammar-nai-ka-no-uchi-ni.jhtml: Note the use with suddenness and tweak the English.

grammar/grammar-ni-atatte.jhtml: Note that the expression is not used with events of a negative nature.

grammar/grammar-sai.jhtml: Enhance the description.

grammar/grammar-totan.jhtml: Note the use with abrupt actions and events, and its unexpected nature. Fix the reference.