More detailed updates from KZM-N2G Chpater 1.4.
grammar/grammar-kagiri-conjunction.jhtml: Note that this form is most often used with a potential verb.

grammar/grammar-kara-shite.jhtml: Note that the following clause is often negative.

grammar/grammar-particle-dake.jhtml: Note the form and nature of the verb involved.

grammar/grammar-wo-hajime-toshite.jhtml: Enhance formation and the description.

grammar/grammar-wo-tooshite.jhtml: Note that the period of time involved should be relatively long.