Fix some problems with stray "@" symbols or missing "{{}}".
These errors lead to a bare @ appearing in the HTML source. This might be visible on the page or it might only be visible in a tooltip.

grammar/grammar-akumade-mo.jhtml:
grammar/grammar-to-douji-ni.jhtml: Drop a bare @.

grammar/grammar-kono-time.jhtml: Fix  "}" that should have been "}}" and "{" that should have been "{{" balanced out to produce a mis-rendered sentence.

grammar/grammar-mai-volitional.jhtml:
grammar/grammar-naku.jhtml: Fix formatting in the formation.

grammar/grammar-soko-de-because.jhtml: Add some missing "{{}}" after a few instances of @S1.

grammar/grammar-te-bakari-wa-irarenai.jhtml: @Vte should have been @Vte{{}}.

study-material/progress/doajg.jhtml: Drop a bare @ in a tooltip (and which was therefore normally invisible).