Verify references for NSM-N3G Week 6 Day 6.
grammar/grammar-ato-de.jhtml: Add a further section (with a label) to cover the usage from NSM-N3G.

grammar/grammar-sore-to.jhtml: Mention that "sore kara" has the same meaning. Note that the reference has been checked.

study-material/toc/nsm-n3g-toc.jhtml: Add a link for "sore to" and an anchor to the link for "ato". Tidy the format too.