Add support for (obscure) hiragana "vu".
test-kana.jhtml has always had an entry for "vu", although it has never worked. That entry is now parsed as an invalid hiragana and so either it needs o be removed or support for "vu" needs to be added. Adding support is easy enough.

scripts/HiraganaSupport.rb: Support hiragana "vu".

test/test-kana.jhtml: As there are entries for "n" and "nn" then "n'" should be tested too.