Make katakana support allow a trailing "n" and have it spot more errors.
scripts/KatakanaSupport.rb: Fix a bad comment. Catch a trailing single katakana character and try to deal with it properly.

grammar/grammar-basic-sentence-patterns.jhtml: Fix the name "Santos" which was misisng the required tailing "u".

grammar/grammar-toka-hearsay.jhtml: "pa-txi" should have been "pa-texi": fixed.

kanji-with-a-single-reading.jhtml: A number of trailing "n" characters changed to "n'". (This was done before the katakana support fix so it has been left it although it should not be needed in future).