Verify the DOBJG reference for the "toka" grammar page.
grammar/grammar-particle-toka.jhtml: Add a GRMIDX for "tokaiu". Show a use with palain sentences in the formation. Add labels. Note that one construction applies only to the general case and not to the specific case. Add a section elaborating on the use of "tokaiu" as a quote marker. Note that the reference has been checked.