Fix up the Kanzen Master reference data.
Touching this data causes a full rebuild. As the previous commit does that for the NSM series, take the opportunity to fix up the KZM series.

data/references.data:

For all the KZM entries, add the Japanese variant.
Make sure that all the known N3, N2 and N1 books are catered for.
For N2 and N1 there are five books. For N3, there seem to be four (no Vocabulary book).
Adjust all the KZM books to use "book name N-level" as the tool-tip title format.