Fix compatibility with ruby 2.1.5.
Testing on an ODROID-C2 running debian jessie showed up an issue, as the default ruby version is 2.1.5.

scripts/build-kanji-vocab-list-by-jlpt.rb: Force encoding unless RUBY_VERSION is 1.8 or less.

scripts/convert-librewriter-html-to-jhtml.rb: Same force encoding here to, but this one seems not to be used by the build and so is untested.