study-material/progress/jpod-advanced.jhtml: Fix JPOD Advanced Season 6 columns. Make other seasons more intuitive.
Add in the missing column headings for Season 6.

For other seasons (except season 3), the new Shadowed and timings columns were added *before* the SRS column.
As they were all empty, this was easy to miss. Moving the '<td> </td>' originally intended to be for SRS to before the
Shadowed column makes the tables easier to edit correctly the first time ...