Cover "wo kagiri ni" for KZM N1 G.
An example sentence needs the kanji for "oomisoka", which RTK has as "last day of the month" so ...

data/kanji.data: ... add "last day of the month" as the official name for Heisig #2452 ...

scripts/FastKanji.rb: ... and commit the rebuilt FastKanji ...

grammar/grammar-wo-kagiri-ni.jhtml: New grammar page for "wo kagiri ni" ...
grammar/Makefile: ... which needs to be built.

study-material/toc/kzm-n1g-toc.jhtml: Link to the new grammar page.