Fix a stray "}" character in each of two TOCs.
An extra "}" showed up as a spurious "}" on the TOC page. It was difficult to spot but what was noticed was that the source text colourisation of an href with an actual grammar entry was wrong starting from a specific point in the file. Tracking that down lead to the discovery of the spurious "}" character.

study-material/toc/kzm-n1g-toc.jhtml:
study-material/toc/nsm-n1g-toc.jhtml: Eliminate a single spurious "}" in each TOC.