Add a page to describe the counter "hako".
I heard "futahako" in an episode of "jimi ni suoi koutsu garu" and couldn't find a onsistent definition in the grammar books.
A Maggie Sensei page had a description and that mostly agreed with the brief one in DOIJG.
The inconsistencies were ironed out at tutorial to produce this final page.

grammar/grammar-counter-hako.jhtml: Create the new grammar page ...

grammar/Makefile: ... build it ...

grammar/grammar-counters.jhtml: ... and link to it from the generic counters page.