grammar/grammar-atte-no.jhtml: Fix the explanation.
The meaning of "N1 atte no N2" is that N2 owes its existence to the presence of N1 or "without N1 there would be no N2".
The original (incorrect) explanation was taken almost verbatim from the Udemy N1 course's lesson 5 explanation in English.
That happens to get the meaning exactly the wrong way round. The Udemy *Japanese* explanation is correct.