grammar/grammar-volitional-ni-mo-nai.jhtml: Fix the nuance involved in this grammar point.
The example in MNN Chuukyuu has a poor English translation.

"can't see the data even if I wanted to" suggests "I don't want to see the data, but even if I did I couldn't".
The Japanese actually means "I would like to see the data but I cannot".

The English explanation says "cannot even if I wanted to" but the Japanese tranlates as "I'd like to but I cannot".

KZM N1 G confirms that the Japanese is correct.