Add a grammar page for "to aimatte".
grammar/grammar-to-aimatte.jhtml: The new grammar page ...

study-material/toc/kzm-n1g-toc.jhtml: ... and a link from the TOC ...

grammar/Makefile: ... and added to the list to build.

data/references.data: Add a reference for "A Handbook of Japanese Grammar Patterns for Teachers and Learners".

data/kanji.data:
scripts/FastKanji.rb: Add a new kanji ("depends on") which looked like it would be needed at one stage.

scripts/jp-to-at.rb: When converting a kanji name that has more than one space (e.g. "sort of thing") into a form suitable for @KJ, replace all the spaces with asterisks, not just the first.