Add a new page to track KZM N1 Dokkai progress.
data/kanji.data:
scripts/FastKanji.rb: Add the kanji for "nab" and "metaphor", both of which are used in the KZM N1 Dokkai progress page.

study-material/progress/kzm-n1r.jhtml: This is the progress page itself ...

study-material/progress/Makefile: ... which needs to be built ...

study-material/progress/study-material.jhtml: ... and linekd to.