study-material/progress/kzm-n1r.jhtml: Bring table up to date by marking all completed exercises.
Once a given exercise has been done, its vocab mined and covered in a lesson, clearly mark it to indicate that nothing else needs to be done. Do this by marking both the page number and the exercise designation as "ok", which will currently make them appear with a green background.