scripts/koohii-stories.py: Begin to generate a web page
This should build a page descrbing each kanji and its story and list any parts and list anything this kanji is part of.

Currently the attempt to turn the codes into a page breaks down because various keywords do not match. Sadly scripts/japanese-to-codes.rb drops out on first failure, so I'd rather pause here and then fix that to list all the failures before dropping out. That would at least show me the scope of the problem!