japanese-to-codes.rb should issue all errors and then fail
The original operation of japanese-to-codes.rb was to stop on first failure. This makes it hard to know how much work needs to be done to fix issues in a new piece of work. The problem was that deep in the operation of the @ command processing, an unknown kanji name was found and that caused the code to report the error and exit.

Now the invalid kanji name can be raised as an error and the higher level processing code can remember the error but keep processing for now. Once all processing has finished, the higher level code can raise an exception.

scripts/ProcessingError.rb: define a unique error code.

scripts/KanjiSupport.rb: use the new error code to raise an exception.

scripts/AtCommandSupport.rb: Spot the new exception, record it, issue a message  and carry on. At the end, if the new exception was raised, raise an exception to that effect.