scripts/koohii-stories.py: sort out the mix-up between @KJ{{...}} and <i> ... </i>
In a story items marked as *multiple words* should become <i>multiple words</i> on the web page (i.e. be in italics).
That's easy enough except that the result of the kanji processing (for a known kanji) will be @KJ{{keyword}} and if the
keyword has spaces in then another script that performs a conversion requires '*' in place of spaces so it becomes
@KJ{{two*words}} or similar. Furthermore, kanji are usually represented in the initial story text by (*{kanji}*),
and that's more or less a requirement from koohii.com, so it is non-negotiable.

That means doing the *...* substitution before kanji substitution breaks all kanji represnetations and doing it
afterwards breaks most if not all kaanji that have multi-word keywords.

The solution is to do the kanji substitutions first but temporarily use a different multi-character string rather
than '*' as the keyword space marker, then perform the *...* substitution, and finally swap back the multi-
character string for '*'