Ensure as many daily use kanji are handled as possible
Update as many keywords as possible to ensure that at least the "kanji" column of the prototype web page is as complete as possible. Note that some older keywords are going to be harder to remove as they will have to be updated and the result manually checked to make sure that nothing has changed. That can be done at a later stage!

data/kanji.data: Update keywords as necessary ...

scripts/FastKanji.rb: ... and rebuild the fast-access script