Fix "sword's point" and "cage [old]" kanji
scripts/KanjiSupport.rb: The kanji name handling code did not allow for the fact that some kanji have [ and ] characters in them and also the possessive marker ’ (not to be confused with ').

I've left in some extra debug and suppressed the 'raise' that normally happens on error, just to help make progress towards getting the prototype heisig keywords web page done.