Update Heisig frame #2255 to use the keyword "muster"
The Heisig keyword "muster" refers to 揃, which is found in frame #2255.
It was listed under the keyword "be complete" and nothing else is listed under "muster".

The only occurrence of "muster" in this repo at this time is in data/transitive-intransitive-verbs.data where it is used for the transitive/intransitive pair 揃える ("soroeru") and 揃う ("sorou"). However, this file is not referenced anywhere, so it seems to be impossible for it to introduce an error.

The only other uses in the whole repository are in the kanji.data file, where it has been introduced as a primary keyword (and also happened to appear as a secondary keyword once), and in my_stories.csv, from koohii.com, which lists heisig-style stories under that keyword.