scripts/AtCommandSupport.rb: Cause an exception on error
If an error happens in a called function, return "PROCESSING-ERROR-HERE", which will be treated as an unrecognised token and will therefore cause an exception to be raised.

At the same time comment out the debug_out lines. Even though it is possible to comment out the single line of debug_out() itself, simply calling it happens often enough that there is a perceptible delay when building a large web page, like the prototype web page or the grammar index. It's easy enough to slip them back in on those occasions when they are actually needed.