makerules/rules.make: Break the build if a JHTML file is invalid
The ruby scripts will generally raise an error or exit with a non-zero status and that will cause make to stop building. However, in that event the Makefil also insists on rm'ing the broken output file, but that would return a 0 success status. So, since rm only ever happens on an error, finish with exit 1 in this case, which make will pick up and cause it to abort the build.