scripts/DataKanji.rb: Improve error reporting
At the moment any parsing error in data/kanji.data just leads to an exception.
To try to be more helpful, report the line number and the line contents and the filename for which the error was triggered.
To try to be even more helpful, report all the errors (i.e. keep running until the end ofthe file) and only raise a final exception if any error was seen. This means all the issues will be reported on the first run, which gives the user a chance to fix all issues at once.