scripts/DataKanji.rb: Slightly improve error handling
I'd prefer the script to ignore blank lines, so do that.

As I'm here anyway, I'd like to make sure that the expected number of fields are actually there. On implementing that I found that many lines had 8 fields and many lines had 9 fields, so I've let that go for now as the data file has been that way for a long time. I'll go back and clean up at a later date (probably :-)).