data/kanji.data: Two "primitive" updates
I've been ignoring "duplicate" notices from build-kanji-ruby.rb but I think some of them will be leading to errors in sentences, so I should try to get to the point where there are no warnings at all.
I'll start with "meeting" (which should be "meeting primitive") and "finger" (which should be "finger primitive").
These are both beyong the ~3000 limit for "normal" kanji and mostly are kanji parts.