scripts/DataKanji.rb: Minimal sanity checks on the input data file
Heisig frame numbers should monotonically increase, atlhough gaps are allowed.
This means there is no need to check for duplicates as they cannot happen.

For any character beyond the Heisig range, koohii.com seems to index them using their
Unicode value. So for all these characters, the unicode value and the frame number
should match. Note that the frame # is recorded in decimal and the Unicode value is in hex.