scripts/jp-to-at.rb: Add ability to process a file
This script has always been used to process a single line of text, supplied on the command line.
Add a --file option to allow it to process an entire file.

Add --file option to process a file instead of whatever is on the CLI.
Change furigana markers from "<" and ">" to "{" and "}".