Do not crash when faced with an unknown kanji
scripts/build-kanji-ruby.rb: If faced with an unknown kanji keyword, no data will be available for it and the attempt to initialise a Kanji object will raise an exception. Rather than letting the exception propogate, catch it, issue a warning to STDERR and return a nil, which the rest of the code already knows how to handle.