Ensure that a failed build for koohii and heisig leaves no output
heisig/Makefile: While debugging the issue with fastKanji and its crash when fed an unknown keyword, I noticed that the output HTML file was left behind, empty. This obviously masks the error as a rebuild will not cause the failing build again as the output is already up to dat (albeit empty). Change the Makefile to ensure that on failure the output is removed (and the Makefile returns a failure).