Kill the build if the Kanji Stories page has an unknown kanji
scripts/japanese-to-codes.rb: Add a CLI option for strict kanji processing.

scripts/AtCommandSupport.rb: If strict kanji processing is active, pass that on to the processing layer below.
There's an ugly global here that needs to vanish. Either that or the exception raising behaviour needs to be the default.

scripts/KanjiSupport.rb: Optionally raise an error if an unknown kanji is requested.

heisig/Makefile: Ensure that a processing failure is noticed and reported up the chain.

makerules/super.make: If any of the subdir Makefiles fails, kill the build.