Break the build if unable to build an output html
There are still invalid kanji in the output. Fix the two that are there now and force a build breakage if any more appear.

makerules/rules.make: Ensure that japanese-to-codes.rb (or at least the main invocation) breaks the build if an invalid kanji is used.

grammar/grammar-ni-hikikae.jhtml: "table" isn't a kanji: the intended one is "wind primitive", which admittedly does look like a table.

study-material/toc/hjgp-tl.jhtml: "wazuka" uses the kanji "trifle", which used to be "a wee bit" in a previous edition from koohii.com.