scripts/KanjiSupport.rb: Add support for "?" in kanji keywords
To allow for the kanji keyword "who?" from koohii.com, kanji keywords have to allow "?" as a valid character.
Enhance the regexp to allow for this. (Also moved "-" to the front ... that seems less surprising for a casual reader).
For maintainability, build the error message once and then decide whether to 'raise' or print it.
Add some comments in passing.