scripts/koohii-stories.py: Improve checks for story formatting
One story had an extra }} that was just a mistyping and not intended to be part of an @-command to be parsed. The ruby processing back-end reacted by raising an exception and prining a very long error message that was, in effect, the rest of the source file.

Fixing the ruby is likely to be tricky, but adjusting the python that parses the story text to spot this (and related) condition, seems fairly straightforward.