Update including "last day of the month"
The kanji 晦 appears in 大晦日 (おおみそか), so it is well worth including.

external/my_stories.csv: New story for "last day of the month".

data/kanji.data: Fix up "last day of the month", which has "*" in place of spaces. That was probably always going to work, but it's wrong and just storing up problems for the future, so fix it.

scripts/FastKanji.rb: Rebuild the fast-access script.