scripts/koohii-stories.py: Fix story formatting for frame #17
Story formatting essentially means converting formthe koohii.com style to raw HTML. The result is actually JHTML, which will be post-processed bythe appropriate script.
The problem is that frame #17 has the keyword "I" so when there is an attempt to substitute "i" with "<b>i</b>", any "i" in the story will be substituted.
There is no stray "i" in the story on koohii.com, but there is during this string processing because kanji substitution will have introduced "<i>" and "</i>".
The simplest course of action is simply to undo the resulting mostrosity and restore the original HTML tags.

The same problem would occur if any keyword was just the letter "B", just in a slightly different form. Luckily no such keyword exists.

Keywords "X" and "V" do exist - these are the only other single character keywords but they seem to cause no problems.